requirements.txt
# Web Framework
fastapi==0.110.0
uvicorn[standard]==0.27.0
pydantic==2.6.0
pydantic-settings==2.1.0
jinja2==3.1.2

# Database (SQLite)
SQLAlchemy==2.0.25

# Config & Utils
pyyaml==6.0.1
python-dotenv==1.0.0
python-dateutil==2.9.0

# Data Processing
pandas==2.2.0
numpy==1.26.3

# Async
aiohttp==3.9.3

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Backtesting
backtrader==1.9.78.123
